Change list of members in activity profile to table view
